requirements.txt
pandas==0.25.1
hickle==3.4.5
tqdm==4.36.1
opencv_python==4.1.1.26
scikit_image==0.15.0
mlconfig==0.0.4
visdom==0.1.8.9
torch==1.0.0
torchvision==0.2.0
yacs==0.1.6
numpy==1.17.4
scipy==1.3.1
apex==0.9.10dev
ipython==7.10.0
Pillow==6.2.1
skimage==0.0
